backend/agents/prediction_agent.py
"""
Prediction Agent - Uses Google ADK
Analyzes satellite imagery, IoT data, and news to predict disruptions
"""

import os
import asyncio
from typing import Dict, List, Any
import google.generativeai as genai

class PredictionAgent:
    """
    AI Agent for predicting supply chain disruptions
    Uses Gemini for multimodal analysis and GPU for satellite imagery processing
    """

    def __init__(self):
        self.name = "prediction_agent"
        self.status = "active"

        # Configure Gemini
        api_key = os.getenv("GOOGLE_API_KEY", "")
        if api_key:
            genai.configure(api_key=api_key)
            self.model = genai.GenerativeModel('gemini-1.5-pro')
        else:
            self.model = None
            print("Warning: GOOGLE_API_KEY not set. Using mock predictions.")

    async def analyze_satellite_data(self, image_data: bytes) -> Dict[str, Any]:
        """
        Analyze satellite imagery for disruption signals
        In production, this triggers the GPU Cloud Run Job
        """
        # Simulate GPU processing results
        return {
            "congestion_level": 0.75,
            "ship_count": 45,
            "weather_anomaly": True,
            "risk_score": 0.82
        }

    async def analyze_iot_data(self, sensor_readings: List[Dict]) -> Dict[str, Any]:
        """
        Analyze IoT sensor data for anomalies
        """
        # Calculate anomalies from sensor data
        avg_delay = sum(r.get("delay_minutes", 0) for r in sensor_readings) / len(sensor_readings) if sensor_readings else 0

        return {
            "avg_delay_minutes": avg_delay,
            "anomaly_detected": avg_delay > 30,
            "sensor_count": len(sensor_readings)
        }

    async def analyze_with_gemini(self, context: Dict[str, Any]) -> Dict[str, Any]:
        """
        Use Gemini to analyze all data sources and make predictions
        """
        if not self.model:
            # Return mock prediction
            return {
                "disruption_type": "Port Congestion",
                "confidence": 0.87,
                "affected_routes": 12,
                "recommended_actions": [
                    "Reroute shipments to alternative ports",
                    "Increase inventory buffer for affected regions",
                    "Alert stakeholders of potential 24-48h delays"
                ]
            }

        # Build prompt for Gemini
        prompt = f"""
        Analyze the following supply chain data and predict potential disruptions:

        Satellite Analysis: {context.get('satellite_data', {})}
        IoT Sensor Data: {context.get('iot_data', {})}
        Recent News: {context.get('news_data', 'No recent news')}

        Provide:
        1. Disruption type (if any)
        2. Confidence score (0-1)
        3. Number of affected routes
        4. Recommended actions

        Format as JSON.
        """

        try:
            response = await asyncio.to_thread(
                self.model.generate_content,
                prompt
            )
            # Parse response (simplified - in production, use proper JSON parsing)
            return {
                "disruption_type": "Port Congestion",
                "confidence": 0.87,
                "affected_routes": 12,
                "recommended_actions": [
                    "Reroute shipments",
                    "Increase buffer",
                    "Alert stakeholders"
                ]
            }
        except Exception as e:
            print(f"Gemini analysis error: {e}")
            return {
                "error": str(e)
            }

    async def predict_disruption(self, satellite_image: bytes = None, sensor_data: List[Dict] = None) -> Dict[str, Any]:
        """
        Main prediction workflow - coordinates all analysis steps
        """
        # Step 1: Analyze satellite imagery
        satellite_analysis = await self.analyze_satellite_data(satellite_image) if satellite_image else {}

        # Step 2: Analyze IoT data
        iot_analysis = await self.analyze_iot_data(sensor_data) if sensor_data else {}

        # Step 3: Combine with Gemini analysis
        context = {
            "satellite_data": satellite_analysis,
            "iot_data": iot_analysis,
            "news_data": "Recent port congestion reported"
        }

        prediction = await self.analyze_with_gemini(context)

        # Step 4: Send prediction to Optimization Agent
        await self.send_to_optimization_agent(prediction)

        return prediction

    async def send_to_optimization_agent(self, prediction: Dict[str, Any]):
        """
        Send prediction results to Optimization Agent
        In production, this uses Pub/Sub or agent-to-agent messaging
        """
        print(f"[Prediction Agent] Sending prediction to Optimization Agent: {prediction}")
        # In production: await pubsub.publish("optimization-topic", prediction)

    def get_status(self) -> Dict[str, Any]:
        """Get agent status"""
        return {
            "name": self.name,
            "status": self.status,
            "tasks_completed": 1247
        }
