backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
google-cloud-aiplatform==1.42.1
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
google-generativeai==0.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
redis==5.0.1
